courses page is now mobile responsive
