+++ react-vite/src/components/AllCoursesPage/AllCourses.css
@@ -1,6 +1,35 @@
-.cm-img {
-    height: 250px;
-    width: 250px;
-    border: solid yellow;
+.cm-container{
+    max-width: 900px;
+    display: grid;
+    grid-template-columns: 1fr 1fr ;
+    grid-gap: 20px;
+    margin: 0 auto;
+    justify-items: center;
+}
+
+.cm-card {
+    height: 350px;
+    width: 350px;
+    border: solid yellow 1px;
     object-fit:contain;
+    background-image: url("https://firebasestorage.googleapis.com/v0/b/captains-log-13590.appspot.com/o/route_log.png?alt=media&token=fdb9ac91-8503-42be-ab1e-b262d910963b");
+    background-size: cover;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    color: rgb(10, 10, 10);
+}
+
+
+@media (min-width: 500px) and (max-width: 700px) {
+
+    .cm-container{
+        max-width: 900px;
+        display: grid;
+        grid-template-columns: 1fr ;
+        grid-gap: 20px;
+        margin: 0 auto;
+        justify-items: center;
+    }
+
 }
